system-test: correct python re.MULTILINE usage
Python re.M flag extends the matching rules for '^' and '$', but it
doesn't mean '.*' will match across the newline char '\n'. For that,
re.DOTALL should be used instead. And since the '*', '+', and '?'
qualifiers are all greedy, and additional '?' should be appended to
prevent it from matching across multiple matches.

Signed-off-by: You-Sheng Yang <vicamo@gmail.com>